+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- API /surveys/list and /surveys/number return error when linked server is "".
+
 [1.19.3] - 2015-10-21
 ----------------------
 ##### Fixed